Change pixel type int16 to int64
Allows large images to be marked with higher values.
That fixes the bugs when "trace" can't find the end node due to integer limit of the x,y address.

+++ sknw/sknw.py
@@ -127,7 +127,7 @@ def mark_node(ske):
     return buf
     
 def build_sknw(ske, buf_size=131072, multi=False, iso=True, ring=True, full=True):
-    buf = np.pad(ske, (1,1), mode='constant').astype(np.uint16)
+    buf = np.pad(ske, (1,1), mode='constant').astype(np.int64)
     nbs = neighbors(buf.shape)
     acc = np.cumprod((1,)+buf.shape[::-1][:-1])[::-1]
     mark(buf, nbs)
